Fix failing case in publicapi test suite
 - Add `ginkgo.report` to .gitignore
 - Introduce `GetTestHandler` and `CheckResponse` functions in middleware tests
 - Remove duplicated test utility functions from `publicapiserver_suite_test.go`
 - Add missing stub routes for eventgenerator
 - Adjust logging in `TLSReloadTransport` to exclude request data
